pyproject.toml
[project]
name = "claude-aso"
version = "1.0.0"
description = "Comprehensive App Store Optimization skill for Claude Code"
requires-python = ">=3.12"
license = "MIT"
readme = "README.md"

dependencies = [
    "beautifulsoup4>=4.12.0,<5.0.0",
    "requests>=2.32.0,<3.0.0",
    "lxml>=5.0.0,<7.0.0",
    "urllib3>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
visual = [
    "playwright>=1.50.0,<2.0.0",
    "Pillow>=11.0.0,<14.0.0",
]
dev = [
    "pytest>=8.3.0,<9.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
